sqlbase: move encoding functions to their own file
Release note: None
